+++ src/main/java/roomescape/dto/request/ThemeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.request;
 
 import roomescape.exception.InvalidInputException;
