+++ src/main/java/roomescape/dto/request/TimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
