+++ src/main/java/roomescape/dto/response/ThemeResponse.java
@@ -1,0 +1,19 @@
+package roomescape.dto.response;
+
+import roomescape.domain.Theme;
+
+public record ThemeResponse(
+    Long id,
+    String name,
+    String description,
+    String thumbnail) {
+
+    public static ThemeResponse from(Theme theme) {
+        return new ThemeResponse(
+            theme.getId(),
+            theme.getName(),
+            theme.getDescription(),
+            theme.getThumbnail()
+        );
+    }
+}
