+++ src/test/java/roomescape/dto/ReservationRequestTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.time.LocalDate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.dto.request.ReservationRequest;
 import roomescape.exception.InvalidInputException;
 
 public class ReservationRequestTest {
