+++ src/test/java/roomescape/dto/TimeRequestTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.dto.request.TimeRequest;
 import roomescape.exception.InvalidInputException;
 
 public class TimeRequestTest {
